Extract native session history and transfer boundary
