+++ packages/agent-native-bridge/agent-native-bridge.cabal
@@ -102,6 +102,8 @@ foreign-library haskell-agent-bridge
                     , Agent.CLI.MacOS.LearnedSkillsBridge
                     , Agent.CLI.MacOS.RepositoryInput
                     , Agent.CLI.MacOS.ResourceAdmin
+                    , Agent.CLI.MacOS.SessionTransferBridge
+                    , Agent.CLI.MacOS.SessionTransferBridge
     c-sources:        cbits/HaskellAgentBridgeRuntime.c
     include-dirs:     include
     includes:         HaskellAgentBridge.h
